Re-implement the old SFunArray
The implementation of SFunArray has changed in the new version
of SBV. Even though the new implementation is better performance-wise,
we still opt for the old one since it supports concrete simulation by
using concrete values instead of the symbolic ones. The new one makes
the values explicitly symbolic even if they are literal constants.